Fixing database_script and adding log file to it
This commit fixes the database_script.py file and adds a log file to it.
Database_script wasn't called after the previous changes in this file.
Now it is called and it works properly.
Log file is added to the project. It logs added and desactivated animals.
It is located in shelters/logs/log.txt